+++ plugins/inputs/opentelemetry/collector/storage.go
@@ -130,7 +130,7 @@ func (ss *SpansStorage) feedAll() {
 
 	if metrics := ss.GetDKMetric(); len(metrics) > 0 {
 		pts := makePoints(metrics)
-		err := dkio.Feed(inputName, datakit.Metric, pts, &dkio.Option{HighFreq: true})
+		err := dkio.Feed(inputName, datakit.Metric, pts, nil)
 		if err != nil {
 			log.Errorf("feed to io error=%v", err)
 		}