+++ plugins/inputs/rum/input_test.go
@@ -27,3 +27,54 @@ func TestLimitReader_Close(t *testing.T) {
 	testutil.Equals(t, 10, len(c))
 	testutil.Equals(t, true, errors.Is(err, errLimitReader))
 }
+
+func TestLookupCDNName(t *testing.T) {
+	ok := isDomainName("xxxxx.xxxxxx")
+	testutil.Assert(t, ok, "")
+
+	_, _, err := lookupCDNName("xxxxx.xxxxxx")
+	t.Log(err)
+	testutil.NotOk(t, err, "")
+
+	{
+		cname, cdn, err := lookupCDNName("www-static.qbox.me")
+		t.Log(cname, cdn)
+		testutil.Ok(t, err)
+	}
+
+	{
+		cname, cdn, err := lookupCDNName("tbcache.com")
+		t.Log(cname, cdn)
+		testutil.Ok(t, err)
+	}
+
+	{
+		cname, cdn, err := lookupCDNName("go.dev")
+		t.Log(cname, cdn, err)
+		testutil.NotOk(t, err, "")
+	}
+
+	{
+		cname, cdn, err := lookupCDNName("ucc.alicdn.com")
+		t.Log(cname, cdn)
+		testutil.Ok(t, err)
+	}
+
+	{
+		cname, cdn, err := lookupCDNName("static.guance.com")
+		t.Log(cname, cdn)
+		testutil.Ok(t, err)
+	}
+
+	{
+		cname, cdn, err := lookupCDNName("res.vmallres.com")
+		t.Log(cname, cdn)
+		testutil.Ok(t, err)
+	}
+
+	{
+		cname, cdn, err := lookupCDNName("cdn.cnbj1.fds.api.mi-img.com")
+		t.Log(cname, cdn)
+		testutil.Ok(t, err)
+	}
+}